Code re-org.
All code in their respective connection file. No functionality change
