+++ src/Makefile.am
@@ -41,8 +41,7 @@ dynomite_SOURCES =			                          \
         dyn_dnode_msg.c dyn_dnode_msg.h                           \
         dyn_dnode_peer.c  dyn_dnode_peer.h                        \
         dyn_dnode_request.c                                       \
-        dyn_dnode_response.c                                      \
-        dyn_dnode_server.c dyn_dnode_server.h                     \
+        dyn_dnode_proxy.c dyn_dnode_proxy.h                     \
         dyn_histogram.c dyn_histogram.h                           \
         dyn_server.c dyn_server.h		                          \
         dyn_proxy.c dyn_proxy.h		                              \
@@ -87,8 +86,7 @@ test_SOURCES =                                                \
         dyn_dnode_msg.c dyn_dnode_msg.h                           \
         dyn_dnode_peer.c  dyn_dnode_peer.h                        \
         dyn_dnode_request.c                                       \
-        dyn_dnode_response.c                                      \
-        dyn_dnode_server.c dyn_dnode_server.h                     \
+        dyn_dnode_proxy.c dyn_dnode_proxy.h                     \
         dyn_histogram.c dyn_histogram.h                           \
         dyn_server.c dyn_server.h                                 \
         dyn_proxy.c dyn_proxy.h                                   \
