+++ src/dyn_dnode_client.h
@@ -9,9 +9,6 @@
 #ifndef _DYN_DNODE_CLIENT_H_
 #define _DYN_DNODE_CLIENT_H_
 
-bool dnode_client_active(struct conn *conn);
-void dnode_client_ref(struct conn *conn, void *owner);
-void dnode_client_unref(struct conn *conn);
-void dnode_client_close(struct context *ctx, struct conn *conn);
+void init_dnode_client_conn(struct conn *conn);
 
 #endif
