+++ src/dyn_dnode_peer.c
@@ -14,7 +14,12 @@
 #include "dyn_token.h"
 
 
-void
+bool is_same_dc(struct server_pool *sp, struct server *peer_node)
+{
+    return string_compare(&sp->dc, &peer_node->dc) == 0;
+}
+
+static void
 dnode_peer_ref(struct conn *conn, void *owner)
 {
     struct server *peer = owner;
@@ -37,7 +42,7 @@ dnode_peer_ref(struct conn *conn, void *owner)
     }
 }
 
-void
+static void
 dnode_peer_unref(struct conn *conn)
 {
     struct server *peer;
@@ -81,7 +86,7 @@ dnode_peer_timeout(struct msg *msg, struct conn *conn)
     return pool->timeout + additional_timeout;
 }
 
-bool
+static bool
 dnode_peer_active(struct conn *conn)
 {
     ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
@@ -162,22 +167,30 @@ dnode_peer_add_local(struct server_pool *pool, struct server *peer)
     return DN_OK;
 }
 
-rstatus_t
-dnode_peer_init(struct array *server_pool, struct context *ctx)
+void
+dnode_peer_deinit(struct array *nodes)
 {
-    rstatus_t status;
+    uint32_t i, nnode;
 
-    status = array_each(server_pool, dnode_peer_each_pool_init, ctx);
-    if (status != DN_OK) {
-        server_pool_deinit(server_pool);
-        return status;
+    for (i = 0, nnode = array_n(nodes); i < nnode; i++) {
+        struct server *s;
+
+        s = array_pop(nodes);
+        IGNORE_RET_VAL(s);
+        ASSERT(TAILQ_EMPTY(&s->s_conn_q) && s->ns_conn_q == 0);
     }
+    array_deinit(nodes);
+}
 
-    return DN_OK;
+static rstatus_t
+dnode_peer_pool_run(struct server_pool *pool)
+{
+    ASSERT(array_n(&pool->peers) != 0);
+    return vnode_update(pool);
 }
 
 
-rstatus_t
+static rstatus_t
 dnode_peer_each_pool_init(void *elem, void *context)
 {
     struct server_pool *sp = (struct server_pool *) elem;
@@ -276,19 +289,18 @@ dnode_peer_each_pool_init(void *elem, void *context)
     return DN_OK;
 }
 
-void
-dnode_peer_deinit(struct array *nodes)
+rstatus_t
+dnode_peer_init(struct array *server_pool, struct context *ctx)
 {
-    uint32_t i, nnode;
+    rstatus_t status;
 
-    for (i = 0, nnode = array_n(nodes); i < nnode; i++) {
-        struct server *s;
-
-        s = array_pop(nodes);
-        IGNORE_RET_VAL(s);
-        ASSERT(TAILQ_EMPTY(&s->s_conn_q) && s->ns_conn_q == 0);
+    status = array_each(server_pool, dnode_peer_each_pool_init, ctx);
+    if (status != DN_OK) {
+        server_pool_deinit(server_pool);
+        return status;
     }
-    array_deinit(nodes);
+
+    return DN_OK;
 }
 
 
@@ -321,7 +333,7 @@ is_conn_secured(struct server_pool *sp, struct server *peer_node)
     return false;
 }
 
-struct conn *
+static struct conn *
 dnode_peer_conn(struct server *server)
 {
     struct server_pool *pool;
@@ -354,57 +366,139 @@ dnode_peer_conn(struct server *server)
     return conn;
 }
 
-
 static rstatus_t
-dnode_peer_each_preconnect(void *elem, void *data)
+dnode_peer_connect(struct context *ctx, struct server *server, struct conn *conn)
 {
     rstatus_t status;
-    struct server *peer;
-    struct server_pool *sp;
-    struct conn *conn;
 
-    peer = elem;
-    sp = peer->owner;
+    if (log_loggable(LOG_VERB)) {
+        log_debug(LOG_VERB, "dnode_peer_connect dyn: connect to peer '%.*s'", server->pname.len,
+                server->pname.data);
+    }
 
-    if (peer->is_local)  //don't bother to connect if it is a self-connection
+    if (ctx->admin_opt > 0)
         return DN_OK;
 
-    conn = dnode_peer_conn(peer);
-    if (conn == NULL) {
-        return DN_ENOMEM;
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    if (conn->sd > 0) {
+        /* already connected on peer connection */
+        return DN_OK;
     }
 
-    status = dnode_peer_connect(sp->ctx, peer, conn);
+    conn->sd = socket(conn->family, SOCK_STREAM, 0);
+    if (conn->sd < 0) {
+        log_error("dyn: socket for peer '%.*s' failed: %s", server->pname.len,
+                server->pname.data, strerror(errno));
+        status = DN_ERROR;
+        goto error;
+    }
+    log_debug(LOG_WARN, "dnode: connected to peer '%.*s' on p %d", server->pname.len,
+            server->pname.data, conn->sd);
+
+
+    status = dn_set_nonblocking(conn->sd);
     if (status != DN_OK) {
-        log_warn("dyn: connect to peer '%.*s' failed, ignored: %s",
-                peer->pname.len, peer->pname.data, strerror(errno));
-        dnode_peer_close(sp->ctx, conn);
+        log_error("dyn: set nonblock on s %d for peer '%.*s' failed: %s",
+                conn->sd,  server->pname.len, server->pname.data,
+                strerror(errno));
+        goto error;
     }
 
+
+    if (server->pname.data[0] != '/') {
+        status = dn_set_tcpnodelay(conn->sd);
+        if (status != DN_OK) {
+            log_warn("dyn: set tcpnodelay on s %d for peer '%.*s' failed, ignored: %s",
+                    conn->sd, server->pname.len, server->pname.data,
+                    strerror(errno));
+        }
+    }
+
+    status = event_add_conn(ctx->evb, conn);
+    if (status != DN_OK) {
+        log_error("dyn: event add conn s %d for peer '%.*s' failed: %s",
+                conn->sd, server->pname.len, server->pname.data,
+                strerror(errno));
+        goto error;
+    }
+
+    ASSERT(!conn->connecting && !conn->connected);
+
+    status = connect(conn->sd, conn->addr, conn->addrlen);
+
+    if (status != DN_OK) {
+        if (errno == EINPROGRESS) {
+            conn->connecting = 1;
+            log_debug(LOG_DEBUG, "dyn: connecting on s %d to peer '%.*s'",
+                    conn->sd, server->pname.len, server->pname.data);
+            return DN_OK;
+        }
+
+        log_error("dyn: connect on s %d to peer '%.*s' failed: %s", conn->sd,
+                server->pname.len, server->pname.data, strerror(errno));
+
+        goto error;
+    }
+
+
+    ASSERT(!conn->connecting);
+    conn->connected = 1;
+    log_debug(LOG_WARN, "dyn: connected on s %d to peer '%.*s'", conn->sd,
+            server->pname.len, server->pname.data);
+
+
     return DN_OK;
+
+    error:
+    conn->err = errno;
+    return status;
 }
 
-static rstatus_t
-dnode_peer_each_disconnect(void *elem, void *data)
+static void
+dnode_peer_ack_err(struct context *ctx, struct conn *conn, struct msg *req)
 {
-    struct server *server;
-    struct server_pool *pool;
-
-    server = elem;
-    pool = server->owner;
-
-    //TODOs: fixe me not to use s_conn_q to distinguish server pool and peer pool
-    while (!TAILQ_EMPTY(&server->s_conn_q)) {
-        struct conn *conn;
-
-        ASSERT(server->ns_conn_q > 0);
-
-        conn = TAILQ_FIRST(&server->s_conn_q);
-        conn_close(pool->ctx, conn);
+    if ((req->swallow && req->noreply) || // no reply
+        (req->swallow && (req->consistency == DC_ONE)) || // dc one
+        (req->swallow && (req->consistency == DC_QUORUM) // remote dc request
+                      && (!conn->same_dc)) ||
+        (req->owner == conn)) // a gossip message that originated on this conn
+    {
+        log_debug(LOG_INFO, "dyn: close s %d swallow req %"PRIu64" len %"PRIu32
+                  " type %d", conn->sd, req->id, req->mlen, req->type);
+        req_put(req);
+        return;
     }
+    struct conn *c_conn = req->owner;
+    // At other connections, these responses would be swallowed.
+    ASSERT_LOG((c_conn->type == CONN_CLIENT) ||
+               (c_conn->type == CONN_DNODE_PEER_CLIENT),
+               "conn:%s c_conn:%s, req %d:%d", conn_get_type_string(conn),
+               conn_get_type_string(c_conn), req->id, req->parent_id);
 
-    return DN_OK;
+    // Create an appropriate response for the request so its propagated up;
+    // This response gets dropped in rsp_make_error anyways. But since this is
+    // an error path its ok with the overhead.
+    struct msg *rsp = msg_get(conn, false, conn->data_store, __FUNCTION__);
+    req->done = 1;
+    rsp->error = req->error = 1;
+    rsp->err = req->err = conn->err;
+    rsp->dyn_error = req->dyn_error = PEER_CONNECTION_REFUSE;
+    rsp->dmsg = dmsg_get();
+    rsp->dmsg->id =  req->id;
+
+    log_warn("dyn: close s %d schedule error for req %u:%u "
+             "len %"PRIu32" type %d from c %d%c %s", conn->sd, req->id, req->parent_id,
+             req->mlen, req->type, c_conn->sd, conn->err ? ':' : ' ',
+             conn->err ? strerror(conn->err): " ");
+    rstatus_t status =
+            conn_handle_response(c_conn, req->parent_id ? req->parent_id : req->id,
+                                 rsp);
+    IGNORE_RET_VAL(status);
+    if (req->swallow)
+        req_put(req);
 }
+
 
 static void
 dnode_peer_failure(struct context *ctx, struct server *server)
@@ -481,88 +575,7 @@ dnode_peer_close_stats(struct context *ctx, struct conn *conn)
 }
 
 
-void
-dnode_peer_attemp_reconnect_or_close(struct context *ctx, struct conn *conn)
-{
-    rstatus_t status;
-
-    if (conn->attempted_reconnect < 30) {
-        status = close(conn->sd);
-        if (status < 0) {
-            log_error("dyn: close s %d failed, ignored: %s", conn->sd, strerror(errno));
-        }
-        conn->sd = -1;
-       if (dnode_peer_connect(ctx, conn->owner, conn) != DN_OK)
-          conn->attempted_reconnect++;
-       else conn->attempted_reconnect = 0;
-    } else {
-        dnode_peer_close(ctx, conn);
-    }
-}
-
 static void
-dnode_peer_close_socket(struct context *ctx, struct conn *conn)
-{
-    rstatus_t status;
-        if (log_loggable(LOG_VERB)) {
-       log_debug(LOG_VERB, "In dnode_peer_close_socket");
-        }
-
-    if (conn != NULL) {
-        status = close(conn->sd);
-        if (status < 0) {
-            log_error("dyn: close s %d failed, ignored: %s", conn->sd, strerror(errno));
-        }
-    }
-
-    conn->dnode_crypto_state = 0;
-    conn->sd = -1;
-}
-
-static void
-dnode_peer_ack_err(struct context *ctx, struct conn *conn, struct msg *req)
-{
-    if ((req->swallow && req->noreply) || // no reply
-        (req->swallow && (req->consistency == DC_ONE)) || // dc one
-        (req->swallow && (req->consistency == DC_QUORUM) // remote dc request
-                      && (!conn->same_dc)) ||
-        (req->owner == conn)) // a gossip message that originated on this conn
-    {
-        log_debug(LOG_INFO, "dyn: close s %d swallow req %"PRIu64" len %"PRIu32
-                  " type %d", conn->sd, req->id, req->mlen, req->type);
-        req_put(req);
-        return;
-    }
-    struct conn *c_conn = req->owner;
-    // At other connections, these responses would be swallowed.
-    ASSERT_LOG((c_conn->type == CONN_CLIENT) ||
-               (c_conn->type == CONN_DNODE_PEER_CLIENT),
-               "conn:%s c_conn:%s, req %d:%d", conn_get_type_string(conn),
-               conn_get_type_string(c_conn), req->id, req->parent_id);
-
-    // Create an appropriate response for the request so its propagated up;
-    // This response gets dropped in rsp_make_error anyways. But since this is
-    // an error path its ok with the overhead.
-    struct msg *rsp = msg_get(conn, false, conn->data_store, __FUNCTION__);
-    req->done = 1;
-    rsp->error = req->error = 1;
-    rsp->err = req->err = conn->err;
-    rsp->dyn_error = req->dyn_error = PEER_CONNECTION_REFUSE;
-    rsp->dmsg = dmsg_get();
-    rsp->dmsg->id =  req->id;
-
-    log_warn("dyn: close s %d schedule error for req %u:%u "
-             "len %"PRIu32" type %d from c %d%c %s", conn->sd, req->id, req->parent_id,
-             req->mlen, req->type, c_conn->sd, conn->err ? ':' : ' ',
-             conn->err ? strerror(conn->err): " ");
-    rstatus_t status =
-            conn_handle_response(c_conn, req->parent_id ? req->parent_id : req->id,
-                                 rsp);
-    IGNORE_RET_VAL(status);
-    if (req->swallow)
-        req_put(req);
-}
-void
 dnode_peer_close(struct context *ctx, struct conn *conn)
 {
     rstatus_t status;
@@ -635,6 +648,78 @@ dnode_peer_close(struct context *ctx, struct conn *conn)
     conn->sd = -1;
 
     conn_put(conn);
+}
+
+
+
+static rstatus_t
+dnode_peer_each_preconnect(void *elem, void *data)
+{
+    rstatus_t status;
+    struct server *peer;
+    struct server_pool *sp;
+    struct conn *conn;
+
+    peer = elem;
+    sp = peer->owner;
+
+    if (peer->is_local)  //don't bother to connect if it is a self-connection
+        return DN_OK;
+
+    conn = dnode_peer_conn(peer);
+    if (conn == NULL) {
+        return DN_ENOMEM;
+    }
+
+    status = dnode_peer_connect(sp->ctx, peer, conn);
+    if (status != DN_OK) {
+        log_warn("dyn: connect to peer '%.*s' failed, ignored: %s",
+                peer->pname.len, peer->pname.data, strerror(errno));
+        dnode_peer_close(sp->ctx, conn);
+    }
+
+    return DN_OK;
+}
+
+static rstatus_t
+dnode_peer_each_disconnect(void *elem, void *data)
+{
+    struct server *server;
+    struct server_pool *pool;
+
+    server = elem;
+    pool = server->owner;
+
+    //TODOs: fixe me not to use s_conn_q to distinguish server pool and peer pool
+    while (!TAILQ_EMPTY(&server->s_conn_q)) {
+        struct conn *conn;
+
+        ASSERT(server->ns_conn_q > 0);
+
+        conn = TAILQ_FIRST(&server->s_conn_q);
+        conn_close(pool->ctx, conn);
+    }
+
+    return DN_OK;
+}
+
+static void
+dnode_peer_close_socket(struct context *ctx, struct conn *conn)
+{
+    rstatus_t status;
+        if (log_loggable(LOG_VERB)) {
+       log_debug(LOG_VERB, "In dnode_peer_close_socket");
+        }
+
+    if (conn != NULL) {
+        status = close(conn->sd);
+        if (status < 0) {
+            log_error("dyn: close s %d failed, ignored: %s", conn->sd, strerror(errno));
+        }
+    }
+
+    conn->dnode_crypto_state = 0;
+    conn->sd = -1;
 }
 
 
@@ -875,29 +960,6 @@ dnode_peer_add(struct server_pool *sp, struct node *node)
 */
 
 rstatus_t
-dnode_peer_remove(void *rmsg)
-{
-    //rstatus_t status;
-    struct ring_msg *msg = rmsg;
-    //struct server_pool *sp = msg->sp;
-    struct node *node = array_get(&msg->nodes, 0);
-    log_debug(LOG_VVERB, "dyn: peer has a removed message '%.*s'", node->name.len, node->name.data);
-    return DN_OK;
-}
-
-/*
-rstatus_t
-dnode_peer_remove(struct server_pool *sp, struct node *node)
-{
-    //rstatus_t status;
-    log_debug(LOG_VVERB, "dyn: peer has a removed message '%.*s'", node->name.len, node->name.data);
-    return DN_OK;
-}
-*/
-
-
-
-rstatus_t
 dnode_peer_replace(void *rmsg)
 {
     //rstatus_t status;
@@ -1014,95 +1076,6 @@ dnode_peer_replace(struct server_pool *sp, struct node *node)
 */
 
 
-rstatus_t
-dnode_peer_connect(struct context *ctx, struct server *server, struct conn *conn)
-{
-    rstatus_t status;
-
-    if (log_loggable(LOG_VERB)) {
-        log_debug(LOG_VERB, "dnode_peer_connect dyn: connect to peer '%.*s'", server->pname.len,
-                server->pname.data);
-    }
-
-    if (ctx->admin_opt > 0)
-        return DN_OK;
-
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-
-    if (conn->sd > 0) {
-        /* already connected on peer connection */
-        return DN_OK;
-    }
-
-    conn->sd = socket(conn->family, SOCK_STREAM, 0);
-    if (conn->sd < 0) {
-        log_error("dyn: socket for peer '%.*s' failed: %s", server->pname.len,
-                server->pname.data, strerror(errno));
-        status = DN_ERROR;
-        goto error;
-    }
-    log_debug(LOG_WARN, "dnode: connected to peer '%.*s' on p %d", server->pname.len,
-            server->pname.data, conn->sd);
-
-
-    status = dn_set_nonblocking(conn->sd);
-    if (status != DN_OK) {
-        log_error("dyn: set nonblock on s %d for peer '%.*s' failed: %s",
-                conn->sd,  server->pname.len, server->pname.data,
-                strerror(errno));
-        goto error;
-    }
-
-
-    if (server->pname.data[0] != '/') {
-        status = dn_set_tcpnodelay(conn->sd);
-        if (status != DN_OK) {
-            log_warn("dyn: set tcpnodelay on s %d for peer '%.*s' failed, ignored: %s",
-                    conn->sd, server->pname.len, server->pname.data,
-                    strerror(errno));
-        }
-    }
-
-    status = event_add_conn(ctx->evb, conn);
-    if (status != DN_OK) {
-        log_error("dyn: event add conn s %d for peer '%.*s' failed: %s",
-                conn->sd, server->pname.len, server->pname.data,
-                strerror(errno));
-        goto error;
-    }
-
-    ASSERT(!conn->connecting && !conn->connected);
-
-    status = connect(conn->sd, conn->addr, conn->addrlen);
-
-    if (status != DN_OK) {
-        if (errno == EINPROGRESS) {
-            conn->connecting = 1;
-            log_debug(LOG_DEBUG, "dyn: connecting on s %d to peer '%.*s'",
-                    conn->sd, server->pname.len, server->pname.data);
-            return DN_OK;
-        }
-
-        log_error("dyn: connect on s %d to peer '%.*s' failed: %s", conn->sd,
-                server->pname.len, server->pname.data, strerror(errno));
-
-        goto error;
-    }
-
-
-    ASSERT(!conn->connecting);
-    conn->connected = 1;
-    log_debug(LOG_WARN, "dyn: connected on s %d to peer '%.*s'", conn->sd,
-            server->pname.len, server->pname.data);
-
-
-    return DN_OK;
-
-    error:
-    conn->err = errno;
-    return status;
-}
-
 void
 dnode_peer_connected(struct context *ctx, struct conn *conn)
 {
@@ -1122,7 +1095,7 @@ dnode_peer_connected(struct context *ctx, struct conn *conn)
         }
 }
 
-void
+static void
 dnode_peer_ok(struct context *ctx, struct conn *conn)
 {
     struct server *server = conn->owner;
@@ -1139,7 +1112,7 @@ dnode_peer_ok(struct context *ctx, struct conn *conn)
     }
 }
 
-rstatus_t
+static rstatus_t
 dnode_peer_pool_update(struct server_pool *pool)
 {
     int64_t now;
@@ -1358,20 +1331,11 @@ dnode_peer_pool_each_disconnect(void *elem, void *data)
     return DN_OK;
 }
 
-void
+static void
 dnode_peer_pool_disconnect(struct context *ctx)
 {
     array_each(&ctx->pool, dnode_peer_pool_each_disconnect, NULL);
 }
-
-rstatus_t
-dnode_peer_pool_run(struct server_pool *pool)
-{
-    ASSERT(array_n(&pool->peers) != 0);
-    return vnode_update(pool);
-}
-
-
 
 void
 dnode_peer_pool_deinit(struct array *server_pool)
@@ -1401,7 +1365,530 @@ dnode_peer_pool_deinit(struct array *server_pool)
 }
 
 
-bool is_same_dc(struct server_pool *sp, struct server *peer_node)
+static struct msg *
+dnode_rsp_recv_next(struct context *ctx, struct conn *conn, bool alloc)
 {
-    return string_compare(&sp->dc, &peer_node->dc) == 0;
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    conn->last_received = time(NULL);
+
+    return rsp_recv_next(ctx, conn, alloc);
+}
+
+static bool
+dnode_rsp_filter(struct context *ctx, struct conn *conn, struct msg *msg)
+{
+    struct msg *pmsg;
+
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    if (msg_empty(msg)) {
+        ASSERT(conn->rmsg == NULL);
+        log_debug(LOG_VERB, "dyn: filter empty rsp %"PRIu64" on s %d", msg->id,
+                conn->sd);
+        rsp_put(msg);
+        return true;
+    }
+
+    pmsg = TAILQ_FIRST(&conn->omsg_q);
+    if (pmsg == NULL) {
+        log_debug(LOG_INFO, "dyn: filter stray rsp %"PRIu64" len %"PRIu32" on s %d noreply %d",
+                msg->id, msg->mlen, conn->sd, msg->noreply);
+        rsp_put(msg);
+        return true;
+    }
+    ASSERT(pmsg->peer == NULL);
+    ASSERT(pmsg->request && !pmsg->done);
+
+    return false;
+}
+
+static void
+dnode_rsp_forward_stats(struct context *ctx, struct server *server, struct msg *msg)
+{
+    ASSERT(!msg->request);
+    stats_pool_incr(ctx, server->owner, peer_responses);
+    stats_pool_incr_by(ctx, server->owner, peer_response_bytes, msg->mlen);
+}
+
+static void
+dnode_rsp_swallow(struct context *ctx, struct conn *peer_conn,
+                  struct msg *req, struct msg *rsp)
+{
+    conn_dequeue_outq(ctx, peer_conn, req);
+    req->done = 1;
+    log_debug(LOG_VERB, "conn %p swallow %p", peer_conn, req);
+    if (rsp) {
+        log_debug(LOG_INFO, "dyn: swallow rsp %"PRIu64" len %"PRIu32" of req "
+                  "%"PRIu64" on s %d", rsp->id, rsp->mlen, req->id,
+                  peer_conn->sd);
+        rsp_put(rsp);
+    }
+    req_put(req);
+}
+
+/* Description: link data from a peer connection to a client-facing connection
+ * peer_conn: a peer connection
+ * msg      : msg with data from the peer connection after parsing
+ */
+static void
+dnode_rsp_forward_match(struct context *ctx, struct conn *peer_conn, struct msg *rsp)
+{
+    rstatus_t status;
+    struct msg *req;
+    struct conn *c_conn;
+
+    req = TAILQ_FIRST(&peer_conn->omsg_q);
+    c_conn = req->owner;
+
+    /* if client consistency is dc_one forward the response from only the
+       local node. Since dyn_dnode_peer is always a remote node, drop the rsp */
+    if (req->consistency == DC_ONE) {
+        if (req->swallow) {
+            dnode_rsp_swallow(ctx, peer_conn, req, rsp);
+            return;
+        }
+        log_warn("req %d:%d with DC_ONE consistency is not being swallowed");
+    }
+
+    /* if client consistency is dc_quorum, forward the response from only the
+       local region/DC. */
+    if ((req->consistency == DC_QUORUM) && !peer_conn->same_dc) {
+        if (req->swallow) {
+            dnode_rsp_swallow(ctx, peer_conn, req, rsp);
+            return;
+        }
+        log_warn("req %d:%d with DC_QUORUM consistency is not being swallowed");
+    }
+
+    log_debug(LOG_DEBUG, "DNODE RSP RECEIVED %s %d dmsg->id %u req %u:%u rsp %u:%u, ",
+              conn_get_type_string(peer_conn),
+              peer_conn->sd, rsp->dmsg->id,
+              req->id, req->parent_id, rsp->id, rsp->parent_id);
+    ASSERT(req != NULL && req->peer == NULL);
+    ASSERT(req->request && !req->done);
+
+    if (log_loggable(LOG_VVERB)) {
+        loga("Dumping content for response:   ");
+        msg_dump(rsp);
+
+        loga("rsp id %d", rsp->id);
+
+        loga("Dumping content for request:");
+        msg_dump(req);
+
+        loga("req id %d", req->id);
+    }
+
+    conn_dequeue_outq(ctx, peer_conn, req);
+    req->done = 1;
+
+    log_info("c_conn:%p %d:%d <-> %d:%d", c_conn, req->id, req->parent_id,
+             rsp->id, rsp->parent_id);
+    /* establish rsp <-> req (response <-> request) link */
+    req->peer = rsp;
+    rsp->peer = req;
+
+    rsp->pre_coalesce(rsp);
+
+    ASSERT_LOG((c_conn->type == CONN_CLIENT) ||
+               (c_conn->type == CONN_DNODE_PEER_CLIENT),
+               "c_conn type %s", conn_get_type_string(c_conn));
+
+    dnode_rsp_forward_stats(ctx, peer_conn->owner, rsp);
+    // c_conn owns respnse now
+    status = conn_handle_response(c_conn, req->parent_id ? req->parent_id : req->id,
+                                  rsp);
+    IGNORE_RET_VAL(status);
+    if (req->swallow) {
+        log_info("swallow request %d:%d", req->id, req->parent_id);
+        req_put(req);
+    }
+}
+
+/* There are chances that the request to the remote peer or its response got dropped.
+ * Hence we may not always receive a response to the request at the head of the FIFO.
+ * Hence what we do is we mark that request as errored and move on the next one
+ * in the outgoing queue. This works since we always have message ids in monotonically
+ * increasing order.
+ */
+static void
+dnode_rsp_forward(struct context *ctx, struct conn *peer_conn, struct msg *rsp)
+{
+    struct msg *req;
+    struct conn *c_conn;
+
+    ASSERT(peer_conn->type == CONN_DNODE_PEER_SERVER);
+
+    /* response from a peer implies that peer is ok and heartbeating */
+    dnode_peer_ok(ctx, peer_conn);
+
+    /* dequeue peer message (request) from peer conn */
+    while (true) {
+        req = TAILQ_FIRST(&peer_conn->omsg_q);
+        log_debug(LOG_VERB, "dnode_rsp_forward entering req %p rsp %p...", req, rsp);
+        c_conn = req->owner;
+
+        if (!peer_conn->same_dc && req->remote_region_send_time) {
+            struct stats *st = ctx->stats;
+            uint64_t delay = dn_usec_now() - req->remote_region_send_time;
+            histo_add(&st->cross_region_histo, delay);
+        }
+
+        if (req->id == rsp->dmsg->id) {
+            dnode_rsp_forward_match(ctx, peer_conn, rsp);
+            return;
+        }
+        // Report a mismatch and try to rectify
+        log_error("MISMATCH: dnode %s %d rsp_dmsg_id %u req %u:%u dnode rsp %u:%u",
+                  conn_get_type_string(peer_conn),
+                  peer_conn->sd, rsp->dmsg->id, req->id, req->parent_id, rsp->id,
+                  rsp->parent_id);
+        if (c_conn && conn_to_ctx(c_conn))
+            stats_pool_incr(conn_to_ctx(c_conn), c_conn->owner,
+                    peer_mismatch_requests);
+
+        // TODO : should you be worried about message id getting wrapped around to 0?
+        if (rsp->dmsg->id < req->id) {
+            // We received a response from the past. This indeed proves out of order
+            // responses. A blunder to the architecture. Log it and drop the response.
+            log_error("MISMATCH: received response from the past. Dropping it");
+            rsp_put(rsp);
+            return;
+        }
+
+        if (req->consistency == DC_ONE) {
+            if (req->swallow) {
+                // swallow the request and move on the next one
+                dnode_rsp_swallow(ctx, peer_conn, req, NULL);
+                continue;
+            }
+            log_warn("req %d:%d with DC_ONE consistency is not being swallowed");
+        }
+
+        if ((req->consistency == DC_QUORUM) && !peer_conn->same_dc) {
+            if (req->swallow) {
+                // swallow the request and move on the next one
+                dnode_rsp_swallow(ctx, peer_conn, req, NULL);
+                continue;
+            }
+            log_warn("req %d:%d with DC_QUORUM consistency is not being swallowed");
+        }
+
+        log_error("MISMATCHED DNODE RSP RECEIVED %s %d dmsg->id %u req %u:%u rsp %u:%u, skipping....",
+                  conn_get_type_string(peer_conn),
+                 peer_conn->sd, rsp->dmsg->id,
+                 req->id, req->parent_id, rsp->id, rsp->parent_id);
+        ASSERT(req != NULL && req->peer == NULL);
+        ASSERT(req->request && !req->done);
+
+        if (log_loggable(LOG_VVERB)) {
+            loga("skipping req:   ");
+            msg_dump(req);
+        }
+
+
+        conn_dequeue_outq(ctx, peer_conn, req);
+        req->done = 1;
+
+        // Create an appropriate response for the request so its propagated up;
+        struct msg *err_rsp = msg_get(peer_conn, false, peer_conn->data_store,
+                                      __FUNCTION__);
+        err_rsp->error = req->error = 1;
+        err_rsp->err = req->err = BAD_FORMAT;
+        err_rsp->dyn_error = req->dyn_error = BAD_FORMAT;
+        err_rsp->dmsg = dmsg_get();
+        err_rsp->dmsg->id = req->id;
+        log_debug(LOG_VERB, "%p <-> %p", req, err_rsp);
+        /* establish err_rsp <-> req (response <-> request) link */
+        req->peer = err_rsp;
+        err_rsp->peer = req;
+
+        log_error("Peer connection s %d skipping request %u:%u, dummy err_rsp %u:%u",
+                 peer_conn->sd, req->id, req->parent_id, err_rsp->id, err_rsp->parent_id);
+        rstatus_t status =
+            conn_handle_response(c_conn, req->parent_id ? req->parent_id : req->id,
+                                err_rsp);
+        IGNORE_RET_VAL(status);
+        if (req->swallow) {
+                log_debug(LOG_INFO, "swallow request %d:%d", req->id, req->parent_id);
+            req_put(req);
+        }
+    }
+}
+
+static void
+dnode_rsp_recv_done(struct context *ctx, struct conn *conn,
+                    struct msg *msg, struct msg *nmsg)
+{
+    log_debug(LOG_VERB, "dnode_rsp_recv_done entering ...");
+
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+    ASSERT(msg != NULL && conn->rmsg == msg);
+    ASSERT(!msg->request);
+    ASSERT(msg->owner == conn);
+    ASSERT(nmsg == NULL || !nmsg->request);
+
+    if (log_loggable(LOG_VVERB)) {
+       loga("Dumping content for msg:   ");
+       msg_dump(msg);
+
+       if (nmsg != NULL) {
+          loga("Dumping content for nmsg :");
+          msg_dump(nmsg);
+       }
+    }
+
+    /* enqueue next message (response), if any */
+    conn->rmsg = nmsg;
+
+    if (dnode_rsp_filter(ctx, conn, msg)) {
+        return;
+    }
+    dnode_rsp_forward(ctx, conn, msg);
+}
+
+
+
+//TODOs: fix this in using dmsg_write with encrypted msgs
+//         It is not in use now.
+/*
+void
+dnode_rsp_gos_syn(struct context *ctx, struct conn *p_conn, struct msg *msg)
+{
+    rstatus_t status;
+    struct msg *pmsg;
+
+    //ASSERT(p_conn->type == CONN_DNODE_PEER_CLIENT);
+
+    //add messsage
+    struct mbuf *nbuf = mbuf_get();
+    if (nbuf == NULL) {
+        log_debug(LOG_ERR, "Error happened in calling mbuf_get");
+        return;  //TODOs: need to address this further
+    }
+
+    msg->done = 1;
+
+    //TODOs: need to free the old msg object
+    pmsg = msg_get(p_conn, 0, msg->redis);
+    if (pmsg == NULL) {
+        mbuf_put(nbuf);
+        return;
+    }
+
+    pmsg->done = 1;
+    // establish msg <-> pmsg (response <-> request) link
+    msg->peer = pmsg;
+    pmsg->peer = msg;
+    pmsg->pre_coalesce(pmsg);
+    pmsg->owner = p_conn;
+
+    //dyn message's meta data
+    uint64_t msg_id = msg->dmsg->id;
+    uint8_t type = GOSSIP_SYN_REPLY;
+    struct string data = string("SYN_REPLY_OK");
+
+    dmsg_write(nbuf, msg_id, type, p_conn, 0);
+    mbuf_insert(&pmsg->mhdr, nbuf);
+
+    //dnode_rsp_recv_done(ctx, p_conn, msg, pmsg);
+    //should we do this?
+    //conn_dequeue_outq(ctx, s_conn, pmsg);
+
+
+
+     //p_conn->enqueue_outq(ctx, p_conn, pmsg);
+     //if (TAILQ_FIRST(&p_conn->omsg_q) != NULL && req_done(p_conn, TAILQ_FIRST(&p_conn->omsg_q))) {
+     //   status = event_add_out(ctx->evb, p_conn);
+     //   if (status != DN_OK) {
+     //      p_conn->err = errno;
+     //   }
+     //}
+
+
+    if (TAILQ_FIRST(&p_conn->omsg_q) != NULL && req_done(p_conn, TAILQ_FIRST(&p_conn->omsg_q))) {
+        status = event_add_out(ctx->evb, p_conn);
+        if (status != DN_OK) {
+            p_conn->err = errno;
+        }
+    }
+
+    //dnode_rsp_forward_stats(ctx, s_conn->owner, msg);
+}
+
+*/
+static struct msg *
+dnode_req_send_next(struct context *ctx, struct conn *conn)
+{
+    rstatus_t status;
+
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    uint32_t now = time(NULL);
+    //throttling the sending traffics here
+    if (!conn->same_dc) {
+        if (conn->last_sent != 0) {
+            uint32_t elapsed_time = now - conn->last_sent;
+            uint32_t earned_tokens = elapsed_time * msgs_per_sec();
+            conn->avail_tokens = (conn->avail_tokens + earned_tokens) < msgs_per_sec()?
+                    conn->avail_tokens + earned_tokens : msgs_per_sec();
+
+        }
+
+        conn->last_sent = now;
+        if (conn->avail_tokens > 0) {
+            conn->avail_tokens--;
+            return req_send_next(ctx, conn);
+        }
+
+        //requeue
+        status = event_add_out(ctx->evb, conn);
+        IGNORE_RET_VAL(status);
+
+        return NULL;
+    }
+
+    conn->last_sent = now;
+    return req_send_next(ctx, conn);
+}
+
+static void
+dnode_req_send_done(struct context *ctx, struct conn *conn, struct msg *msg)
+{
+    if (log_loggable(LOG_DEBUG)) {
+       log_debug(LOG_VERB, "dnode_req_send_done entering!!!");
+    }
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+    // crashes because dmsg is NULL :(
+    /*log_debug(LOG_DEBUG, "DNODE REQ SEND %s %d dmsg->id %u",
+              conn_get_type_string(conn), conn->sd, msg->dmsg->id);*/
+
+    if (!conn->same_dc)
+        msg->remote_region_send_time = dn_usec_now();
+    req_send_done(ctx, conn, msg);
+}
+
+static void
+dnode_req_peer_enqueue_imsgq(struct context *ctx, struct conn *conn, struct msg *msg)
+{
+    ASSERT(msg->request);
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    log_debug(LOG_VERB, "conn %p enqueue inq %d:%d calling req_server_enqueue_imsgq",
+              conn, msg->id, msg->parent_id);
+    if (!msg->noreply) {
+        msg_tmo_insert(msg, conn);
+    }
+    TAILQ_INSERT_TAIL(&conn->imsg_q, msg, s_tqe);
+    log_debug(LOG_VERB, "conn %p enqueue inq %d:%d", conn, msg->id, msg->parent_id);
+
+    conn->imsg_count++;
+    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
+    if (conn->same_dc) {
+        histo_add(&ctx->stats->peer_in_queue, conn->imsg_count);
+        stats_pool_incr(ctx, pool, peer_in_queue);
+        stats_pool_incr_by(ctx, pool, peer_in_queue_bytes, msg->mlen);
+    } else {
+        histo_add(&ctx->stats->remote_peer_in_queue, conn->imsg_count);
+        stats_pool_incr(ctx, pool, remote_peer_in_queue);
+        stats_pool_incr_by(ctx, pool, remote_peer_in_queue_bytes, msg->mlen);
+    }
+
+}
+
+static void
+dnode_req_peer_dequeue_imsgq(struct context *ctx, struct conn *conn, struct msg *msg)
+{
+    ASSERT(msg->request);
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    TAILQ_REMOVE(&conn->imsg_q, msg, s_tqe);
+    log_debug(LOG_VERB, "conn %p dequeue inq %d:%d", conn, msg->id, msg->parent_id);
+
+    conn->imsg_count--;
+    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
+    if (conn->same_dc) {
+        histo_add(&ctx->stats->peer_in_queue, conn->imsg_count);
+        stats_pool_decr(ctx, pool, peer_in_queue);
+        stats_pool_decr_by(ctx, pool, peer_in_queue_bytes, msg->mlen);
+    } else {
+        histo_add(&ctx->stats->remote_peer_in_queue, conn->imsg_count);
+        stats_pool_decr(ctx, pool, remote_peer_in_queue);
+        stats_pool_decr_by(ctx, pool, remote_peer_in_queue_bytes, msg->mlen);
+    }
+}
+
+static void
+dnode_req_peer_enqueue_omsgq(struct context *ctx, struct conn *conn, struct msg *msg)
+{
+    ASSERT(msg->request);
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    TAILQ_INSERT_TAIL(&conn->omsg_q, msg, s_tqe);
+    log_debug(LOG_VERB, "conn %p enqueue outq %d:%d", conn, msg->id, msg->parent_id);
+
+    //use only the 1st pool
+    conn->omsg_count++;
+    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
+    if (conn->same_dc) {
+        histo_add(&ctx->stats->peer_out_queue, conn->omsg_count);
+        stats_pool_incr(ctx, pool, peer_out_queue);
+        stats_pool_incr_by(ctx, pool, peer_out_queue_bytes, msg->mlen);
+    } else {
+        histo_add(&ctx->stats->remote_peer_out_queue, conn->omsg_count);
+        stats_pool_incr(ctx, pool, remote_peer_out_queue);
+        stats_pool_incr_by(ctx, pool, remote_peer_out_queue_bytes, msg->mlen);
+    }
+}
+
+static void
+dnode_req_peer_dequeue_omsgq(struct context *ctx, struct conn *conn, struct msg *msg)
+{
+    ASSERT(msg->request);
+    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
+
+    msg_tmo_delete(msg);
+
+    TAILQ_REMOVE(&conn->omsg_q, msg, s_tqe);
+    log_debug(LOG_VVERB, "conn %p dequeue outq %p", conn, msg);
+
+    //use the 1st pool
+    conn->omsg_count--;
+    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
+    if (conn->same_dc) {
+        histo_add(&ctx->stats->peer_out_queue, conn->omsg_count);
+        stats_pool_decr(ctx, pool, peer_out_queue);
+        stats_pool_decr_by(ctx, pool, peer_out_queue_bytes, msg->mlen);
+    } else {
+        histo_add(&ctx->stats->remote_peer_out_queue, conn->omsg_count);
+        stats_pool_decr(ctx, pool, remote_peer_out_queue);
+        stats_pool_decr_by(ctx, pool, remote_peer_out_queue_bytes, msg->mlen);
+    }
+}
+
+
+struct conn_ops dnode_peer_ops = {
+    msg_recv,
+    dnode_rsp_recv_next,
+    dnode_rsp_recv_done,
+    msg_send,
+    dnode_req_send_next,
+    dnode_req_send_done,
+    dnode_peer_close,
+    dnode_peer_active,
+    dnode_peer_ref,
+    dnode_peer_unref,
+    dnode_req_peer_enqueue_imsgq,
+    dnode_req_peer_dequeue_imsgq,
+    dnode_req_peer_enqueue_omsgq,
+    dnode_req_peer_dequeue_omsgq,
+    conn_cant_handle_response
+};
+
+void
+init_dnode_peer_conn(struct conn *conn)
+{
+    conn->type = CONN_DNODE_PEER_SERVER;
+    conn->ops = &dnode_peer_ops;
 }
