+++ src/dyn_dnode_request.c
@@ -12,172 +12,6 @@
 //static struct string client_request_dyn_msg = string("Client_request");
 static uint64_t peer_msg_id = 0;
 
-void
-dnode_req_peer_enqueue_imsgq(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(msg->request);
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-
-    log_debug(LOG_VERB, "conn %p enqueue inq %d:%d calling req_server_enqueue_imsgq",
-              conn, msg->id, msg->parent_id);
-    if (!msg->noreply) {
-        msg_tmo_insert(msg, conn);
-    }
-    TAILQ_INSERT_TAIL(&conn->imsg_q, msg, s_tqe);
-    log_debug(LOG_VERB, "conn %p enqueue inq %d:%d", conn, msg->id, msg->parent_id);
-
-    conn->imsg_count++;
-    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
-    if (conn->same_dc) {
-        histo_add(&ctx->stats->peer_in_queue, conn->imsg_count);
-        stats_pool_incr(ctx, pool, peer_in_queue);
-        stats_pool_incr_by(ctx, pool, peer_in_queue_bytes, msg->mlen);
-    } else {
-        histo_add(&ctx->stats->remote_peer_in_queue, conn->imsg_count);
-        stats_pool_incr(ctx, pool, remote_peer_in_queue);
-        stats_pool_incr_by(ctx, pool, remote_peer_in_queue_bytes, msg->mlen);
-    }
-
-}
-
-void
-dnode_req_peer_dequeue_imsgq(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(msg->request);
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-
-    TAILQ_REMOVE(&conn->imsg_q, msg, s_tqe);
-    log_debug(LOG_VERB, "conn %p dequeue inq %d:%d", conn, msg->id, msg->parent_id);
-
-    conn->imsg_count--;
-    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
-    if (conn->same_dc) {
-        histo_add(&ctx->stats->peer_in_queue, conn->imsg_count);
-        stats_pool_decr(ctx, pool, peer_in_queue);
-        stats_pool_decr_by(ctx, pool, peer_in_queue_bytes, msg->mlen);
-    } else {
-        histo_add(&ctx->stats->remote_peer_in_queue, conn->imsg_count);
-        stats_pool_decr(ctx, pool, remote_peer_in_queue);
-        stats_pool_decr_by(ctx, pool, remote_peer_in_queue_bytes, msg->mlen);
-    }
-}
-
-void
-dnode_req_client_enqueue_omsgq(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(msg->request);
-    ASSERT(conn->type == CONN_DNODE_PEER_CLIENT);
-
-    log_debug(LOG_VERB, "conn %p enqueue outq %p", conn, msg);
-    TAILQ_INSERT_TAIL(&conn->omsg_q, msg, c_tqe);
-
-    //use only the 1st pool
-    conn->omsg_count++;
-    histo_add(&ctx->stats->dnode_client_out_queue, conn->omsg_count);
-    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
-    stats_pool_incr(ctx, pool, dnode_client_out_queue);
-    stats_pool_incr_by(ctx, pool, dnode_client_out_queue_bytes, msg->mlen);
-}
-
-void
-dnode_req_peer_enqueue_omsgq(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(msg->request);
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-
-    TAILQ_INSERT_TAIL(&conn->omsg_q, msg, s_tqe);
-    log_debug(LOG_VERB, "conn %p enqueue outq %d:%d", conn, msg->id, msg->parent_id);
-
-    //use only the 1st pool
-    conn->omsg_count++;
-    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
-    if (conn->same_dc) {
-        histo_add(&ctx->stats->peer_out_queue, conn->omsg_count);
-        stats_pool_incr(ctx, pool, peer_out_queue);
-        stats_pool_incr_by(ctx, pool, peer_out_queue_bytes, msg->mlen);
-    } else {
-        histo_add(&ctx->stats->remote_peer_out_queue, conn->omsg_count);
-        stats_pool_incr(ctx, pool, remote_peer_out_queue);
-        stats_pool_incr_by(ctx, pool, remote_peer_out_queue_bytes, msg->mlen);
-    }
-}
-
-void
-dnode_req_client_dequeue_omsgq(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(msg->request);
-    ASSERT(conn->type == CONN_DNODE_PEER_CLIENT);
-
-    TAILQ_REMOVE(&conn->omsg_q, msg, c_tqe);
-    log_debug(LOG_VERB, "conn %p dequeue outq %p", conn, msg);
-
-    //use the 1st pool
-    conn->omsg_count--;
-    histo_add(&ctx->stats->dnode_client_out_queue, conn->omsg_count);
-    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
-    stats_pool_decr(ctx, pool, dnode_client_out_queue);
-    stats_pool_decr_by(ctx, pool, dnode_client_out_queue_bytes, msg->mlen);
-}
-
-void
-dnode_req_peer_dequeue_omsgq(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(msg->request);
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-
-    msg_tmo_delete(msg);
-
-    TAILQ_REMOVE(&conn->omsg_q, msg, s_tqe);
-    log_debug(LOG_VVERB, "conn %p dequeue outq %p", conn, msg);
-
-    //use the 1st pool
-    conn->omsg_count--;
-    struct server_pool *pool = (struct server_pool *) array_get(&ctx->pool, 0);
-    if (conn->same_dc) {
-        histo_add(&ctx->stats->peer_out_queue, conn->omsg_count);
-        stats_pool_decr(ctx, pool, peer_out_queue);
-        stats_pool_decr_by(ctx, pool, peer_out_queue_bytes, msg->mlen);
-    } else {
-        histo_add(&ctx->stats->remote_peer_out_queue, conn->omsg_count);
-        stats_pool_decr(ctx, pool, remote_peer_out_queue);
-        stats_pool_decr_by(ctx, pool, remote_peer_out_queue_bytes, msg->mlen);
-    }
-}
-
-struct msg *
-dnode_req_recv_next(struct context *ctx, struct conn *conn, bool alloc)
-{
-    ASSERT(conn->type == CONN_DNODE_PEER_CLIENT);
-    return req_recv_next(ctx, conn, alloc);
-}
-
-static bool
-dnode_req_filter(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    ASSERT(conn->type == CONN_DNODE_PEER_CLIENT);
-
-    if (msg_empty(msg)) {
-        ASSERT(conn->rmsg == NULL);
-        if (log_loggable(LOG_VERB)) {
-           log_debug(LOG_VERB, "dyn: filter empty req %"PRIu64" from c %d", msg->id,
-                       conn->sd);
-        }
-        req_put(msg);
-        return true;
-    }
-
-    /* dynomite handler */
-    if (msg->dmsg != NULL) {
-        if (dmsg_process(ctx, conn, msg->dmsg)) {
-            req_put(msg);
-            return true;
-        }
-
-    }
-
-    return false;
-}
-
 static void
 dnode_req_forward_error(struct context *ctx, struct conn *conn, struct msg *msg)
 {
@@ -208,159 +42,6 @@ dnode_req_forward_error(struct context *ctx, struct conn *conn, struct msg *msg)
 
 }
 
-
-static void
-dnode_req_forward(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    struct server_pool *pool;
-    uint8_t *key;
-    uint32_t keylen;
-
-    if (log_loggable(LOG_DEBUG)) {
-       log_debug(LOG_DEBUG, "dnode_req_forward entering ");
-    }
-    log_debug(LOG_DEBUG, "DNODE REQ RECEIVED %s %d dmsg->id %u",
-              conn_get_type_string(conn), conn->sd, msg->dmsg->id);
-
-    ASSERT(conn->type == CONN_DNODE_PEER_CLIENT);
-
-    pool = conn->owner;
-    key = NULL;
-    keylen = 0;
-
-    if (!string_empty(&pool->hash_tag)) {
-        struct string *tag = &pool->hash_tag;
-        uint8_t *tag_start, *tag_end;
-
-        tag_start = dn_strchr(msg->key_start, msg->key_end, tag->data[0]);
-        if (tag_start != NULL) {
-            tag_end = dn_strchr(tag_start + 1, msg->key_end, tag->data[1]);
-            if (tag_end != NULL) {
-                key = tag_start + 1;
-                keylen = (uint32_t)(tag_end - key);
-            }
-        }
-    }
-
-    if (keylen == 0) {
-        key = msg->key_start;
-        keylen = (uint32_t)(msg->key_end - msg->key_start);
-    }
-
-    ASSERT(msg->dmsg != NULL);
-    if (msg->dmsg->type == DMSG_REQ) {
-       local_req_forward(ctx, conn, msg, key, keylen);
-    } else if (msg->dmsg->type == DMSG_REQ_FORWARD) {
-        struct mbuf *orig_mbuf = STAILQ_FIRST(&msg->mhdr);
-        struct datacenter *dc = server_get_dc(pool, &pool->dc);
-        uint32_t rack_cnt = array_n(&dc->racks);
-        uint32_t rack_index;
-        for(rack_index = 0; rack_index < rack_cnt; rack_index++) {
-            struct rack *rack = array_get(&dc->racks, rack_index);
-            //log_debug(LOG_DEBUG, "forwarding to rack  '%.*s'",
-            //            rack->name->len, rack->name->data);
-            struct msg *rack_msg;
-            if (string_compare(rack->name, &pool->rack) == 0 ) {
-                rack_msg = msg;
-            } else {
-                rack_msg = msg_get(conn, msg->request, msg->data_store, __FUNCTION__);
-                if (rack_msg == NULL) {
-                    log_debug(LOG_VERB, "whelp, looks like yer screwed now, buddy. no inter-rack messages for you!");
-                    continue;
-                }
-
-                if (msg_clone(msg, orig_mbuf, rack_msg) != DN_OK) {
-                    msg_put(rack_msg);
-                    continue;
-                }
-                rack_msg->swallow = true;
-            }
-
-            if (log_loggable(LOG_DEBUG)) {
-               log_debug(LOG_DEBUG, "forwarding request from conn '%s' to rack '%.*s' dc '%.*s' ",
-                           dn_unresolve_peer_desc(conn->sd), rack->name->len, rack->name->data, rack->dc->len, rack->dc->data);
-            }
-
-            remote_req_forward(ctx, conn, rack_msg, rack, key, keylen);
-        }
-    }
-}
-
-
-void
-dnode_req_recv_done(struct context *ctx, struct conn *conn,
-                    struct msg *msg, struct msg *nmsg)
-{
-    ASSERT(conn->type == CONN_DNODE_PEER_CLIENT);
-    ASSERT(msg->request);
-    ASSERT(msg->owner == conn);
-    ASSERT(conn->rmsg == msg);
-    ASSERT(nmsg == NULL || nmsg->request);
-
-    /* enqueue next message (request), if any */
-    conn->rmsg = nmsg;
-
-    if (dnode_req_filter(ctx, conn, msg)) {
-        return;
-    }
-
-    log_debug(LOG_VERB, "received msg: %d:%d", msg->id, msg->parent_id);
-    dnode_req_forward(ctx, conn, msg);
-}
-
-
-struct msg *
-dnode_req_send_next(struct context *ctx, struct conn *conn)
-{
-    rstatus_t status;
-
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-
-    uint32_t now = time(NULL);
-    //throttling the sending traffics here
-    if (!conn->same_dc) {
-        if (conn->last_sent != 0) {
-            uint32_t elapsed_time = now - conn->last_sent;
-            uint32_t earned_tokens = elapsed_time * msgs_per_sec();
-            conn->avail_tokens = (conn->avail_tokens + earned_tokens) < msgs_per_sec()?
-                    conn->avail_tokens + earned_tokens : msgs_per_sec();
-
-        }
-
-        conn->last_sent = now;
-        if (conn->avail_tokens > 0) {
-            conn->avail_tokens--;
-            return req_send_next(ctx, conn);
-        }
-
-        //requeue
-        status = event_add_out(ctx->evb, conn);
-        IGNORE_RET_VAL(status);
-
-        return NULL;
-    }
-
-    conn->last_sent = now;
-    return req_send_next(ctx, conn);
-}
-
-void
-dnode_req_send_done(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    if (log_loggable(LOG_DEBUG)) {
-       log_debug(LOG_VERB, "dnode_req_send_done entering!!!");
-    }
-    ASSERT(conn->type == CONN_DNODE_PEER_SERVER);
-    // crashes because dmsg is NULL :(
-    /*log_debug(LOG_DEBUG, "DNODE REQ SEND %s %d dmsg->id %u",
-              conn_get_type_string(conn), conn->sd, msg->dmsg->id);*/
-
-    if (!conn->same_dc)
-        msg->remote_region_send_time = dn_usec_now();
-    req_send_done(ctx, conn, msg);
-}
-
-
 static void
 dnode_peer_req_forward_stats(struct context *ctx, struct server *server, struct msg *msg)
 {
@@ -374,9 +55,10 @@ dnode_peer_req_forward_stats(struct context *ctx, struct server *server, struct 
 
 
 /* Forward a client request over to a peer */
-void dnode_peer_req_forward(struct context *ctx, struct conn *c_conn,
-                            struct conn *p_conn, struct msg *msg,
-                            struct rack *rack, uint8_t *key, uint32_t keylen)
+void
+dnode_peer_req_forward(struct context *ctx, struct conn *c_conn,
+                       struct conn *p_conn, struct msg *msg,
+                       struct rack *rack, uint8_t *key, uint32_t keylen)
 {
 
     struct server *server = p_conn->owner;
