+++ src/dyn_proxy.h
@@ -25,17 +25,7 @@
 #ifndef _DYN_PROXY_H_
 #define _DYN_PROXY_H_
 
-
-
-void proxy_ref(struct conn *conn, void *owner);
-void proxy_unref(struct conn *conn);
-void proxy_close(struct context *ctx, struct conn *conn);
-
-rstatus_t proxy_each_init(void *elem, void *data);
-rstatus_t proxy_each_deinit(void *elem, void *data);
-
 rstatus_t proxy_init(struct context *ctx);
 void proxy_deinit(struct context *ctx);
-rstatus_t proxy_recv(struct context *ctx, struct conn *conn);
 
 #endif
