+++ src/dyn_response.c
@@ -93,196 +93,6 @@ rsp_make_error(struct context *ctx, struct conn *conn, struct msg *msg)
 }
 
 struct msg *
-rsp_recv_next(struct context *ctx, struct conn *conn, bool alloc)
-{
-    struct msg *msg;
-
-    ASSERT((conn->type == CONN_DNODE_PEER_SERVER) ||
-           (conn->type == CONN_SERVER));
-
-    if (conn->eof) {
-        msg = conn->rmsg;
-
-        if (conn->dyn_mode) {
-            if (conn->non_bytes_recv > MAX_CONN_ALLOWABLE_NON_RECV) {
-                conn->err = EPIPE;
-                return NULL;
-            }
-            conn->eof = 0;
-            return msg;
-        }
-
-        /* server sent eof before sending the entire request */
-        if (msg != NULL) {
-            conn->rmsg = NULL;
-
-            ASSERT(msg->peer == NULL);
-            ASSERT(!msg->request);
-
-            log_error("eof s %d discarding incomplete rsp %"PRIu64" len "
-                      "%"PRIu32"", conn->sd, msg->id, msg->mlen);
-
-            rsp_put(msg);
-        }
-
-        /*
-         * We treat TCP half-close from a server different from how we treat
-         * those from a client. On a FIN from a server, we close the connection
-         * immediately by sending the second FIN even if there were outstanding
-         * or pending requests. This is actually a tricky part in the FA, as
-         * we don't expect this to happen unless the server is misbehaving or
-         * it crashes
-         */
-        conn->done = 1;
-        log_debug(LOG_DEBUG, "s %d active %d is done", conn->sd, conn_active(conn));
-
-        return NULL;
-    }
-
-    msg = conn->rmsg;
-    if (msg != NULL) {
-        ASSERT(!msg->request);
-        return msg;
-    }
-
-    if (!alloc) {
-        return NULL;
-    }
-
-    msg = rsp_get(conn);
-    if (msg != NULL) {
-        conn->rmsg = msg;
-    }
-
-    return msg;
-}
-
-static bool
-server_rsp_filter(struct context *ctx, struct conn *conn, struct msg *msg)
-{
-    struct msg *pmsg;
-
-    ASSERT(conn->type == CONN_SERVER);
-
-    if (msg_empty(msg)) {
-        ASSERT(conn->rmsg == NULL);
-        log_debug(LOG_VERB, "filter empty rsp %"PRIu64" on s %d", msg->id,
-                  conn->sd);
-        rsp_put(msg);
-        return true;
-    }
-
-    pmsg = TAILQ_FIRST(&conn->omsg_q);
-    if (pmsg == NULL) {
-        log_debug(LOG_VERB, "filter stray rsp %"PRIu64" len %"PRIu32" on s %d",
-                  msg->id, msg->mlen, conn->sd);
-        rsp_put(msg);
-        return true;
-    }
-
-    if (pmsg->noreply) {
-         conn_dequeue_outq(ctx, conn, pmsg);
-         req_put(pmsg);
-         rsp_put(msg);
-         return true;
-    }
-
-    ASSERT(pmsg->peer == NULL);
-    ASSERT(pmsg->request && !pmsg->done);
-
-    if (pmsg->swallow) {
-        conn_dequeue_outq(ctx, conn, pmsg);
-        pmsg->done = 1;
-
-        log_debug(LOG_DEBUG, "swallow rsp %"PRIu64" len %"PRIu32" of req "
-                  "%"PRIu64" on s %d", msg->id, msg->mlen, pmsg->id,
-                  conn->sd);
-
-        rsp_put(msg);
-        req_put(pmsg);
-        return true;
-    }
-
-    return false;
-}
-
-static void
-server_rsp_forward_stats(struct context *ctx, struct server *server,
-                         struct msg *msg)
-{
-	ASSERT(!msg->request);
-
-	if (msg->is_read) {
-		stats_server_incr(ctx, server, read_responses);
-		stats_server_incr_by(ctx, server, read_response_bytes, msg->mlen);
-	} else {
-		stats_server_incr(ctx, server, write_responses);
-		stats_server_incr_by(ctx, server, write_response_bytes, msg->mlen);
-	}
-}
-
-static void
-server_rsp_forward(struct context *ctx, struct conn *s_conn, struct msg *rsp)
-{
-    rstatus_t status;
-    struct msg *req;
-    struct conn *c_conn;
-    ASSERT(s_conn->type == CONN_SERVER);
-
-    /* response from server implies that server is ok and heartbeating */
-    server_ok(ctx, s_conn);
-
-    /* dequeue peer message (request) from server */
-    req = TAILQ_FIRST(&s_conn->omsg_q);
-    ASSERT(req != NULL && req->peer == NULL);
-    ASSERT(req->request && !req->done);
-
-    conn_dequeue_outq(ctx, s_conn, req);
-    req->done = 1;
-
-    /* establish rsp <-> req (response <-> request) link */
-    req->peer = rsp;
-    rsp->peer = req;
-
-    rsp->pre_coalesce(rsp);
-
-    c_conn = req->owner;
-    log_info("c_conn %p %d:%d <-> %d:%d", c_conn, req->id, req->parent_id,
-               rsp->id, rsp->parent_id);
-    
-    ASSERT((c_conn->type == CONN_CLIENT) ||
-           (c_conn->type == CONN_DNODE_PEER_CLIENT));
-
-    server_rsp_forward_stats(ctx, s_conn->owner, rsp);
-    // this should really be the message's response handler be doing it
-    if (req_done(c_conn, req)) {
-        // handler owns the response now
-        status = conn_handle_response(c_conn, c_conn->type == CONN_CLIENT ?
-                                      req->id : req->parent_id, rsp);
-        IGNORE_RET_VAL(status);
-     }
-}
-
-void
-server_rsp_recv_done(struct context *ctx, struct conn *conn, struct msg *msg,
-                     struct msg *nmsg)
-{
-    ASSERT(conn->type == CONN_SERVER);
-    ASSERT(msg != NULL && conn->rmsg == msg);
-    ASSERT(!msg->request);
-    ASSERT(msg->owner == conn);
-    ASSERT(nmsg == NULL || !nmsg->request);
-
-    /* enqueue next message (response), if any */
-    conn->rmsg = nmsg;
-
-    if (server_rsp_filter(ctx, conn, msg)) {
-        return;
-    }
-    server_rsp_forward(ctx, conn, msg);
-}
-
-struct msg *
 rsp_send_next(struct context *ctx, struct conn *conn)
 {
     rstatus_t status;
