+++ src/dyn_server.h
@@ -70,30 +70,15 @@
  */
 
 
-void server_ref(struct conn *conn, void *owner);
-void server_unref(struct conn *conn);
 int server_timeout(struct conn *conn);
-bool server_active(struct conn *conn);
 rstatus_t server_init(struct array *server, struct array *conf_server, struct server_pool *sp);
-void server_deinit(struct array *server);
-struct conn *server_conn(struct server *server);
 rstatus_t server_connect(struct context *ctx, struct server *server, struct conn *conn);
-void server_close(struct context *ctx, struct conn *conn);
-void server_connected(struct context *ctx, struct conn *conn);
-void server_ok(struct context *ctx, struct conn *conn);
 
 struct datacenter *server_get_dc(struct server_pool *pool, struct string *dcname);
 struct rack *server_get_rack(struct datacenter *dc, struct string *rackname);
 struct rack *server_get_rack_by_dc_rack(struct server_pool *sp, struct string *rackname, struct string *dcname);
 
-rstatus_t rack_init(struct rack *rack);
-rstatus_t rack_deinit(struct rack *rack);
-rstatus_t rack_destroy(void *elem, void *data);
-
-rstatus_t dc_init(struct datacenter *dc);
-rstatus_t dc_deinit(struct datacenter *dc);
 rstatus_t datacenter_destroy(void *elem, void *data);
-
 
 struct conn *server_pool_conn(struct context *ctx, struct server_pool *pool, uint8_t *key, uint32_t keylen);
 rstatus_t server_pool_run(struct server_pool *pool);
@@ -101,5 +86,6 @@ rstatus_t server_pool_preconnect(struct context *ctx);
 void server_pool_disconnect(struct context *ctx);
 rstatus_t server_pool_init(struct array *server_pool, struct array *conf_pool, struct context *ctx);
 void server_pool_deinit(struct array *server_pool);
+void init_server_conn(struct conn *conn);
 
 #endif
